Create AlphabeticAnagrams.py
A solution to the kata 'Alphabetic Anagrams' by user Azuaron on CodeWars. I had a fun time solving this and working out the optimum solution that satisfied all edge cases, although the most trouble I had was as a result of not using integer division which lead to my answers being slightly off until I narrowed down the problem.